+++ longevity/src/main/scala/longevity/exceptions/subdomain/InvalidPartitionException.scala
@@ -9,7 +9,7 @@ import longevity.subdomain.ptype.PartitionKey
  * @param path the requested property path
  * @param pTypeKey the type of the persistent with the problematic property
  */
-class InvalidPartitionException[P : TypeKey] (val key: PartitionKey[P, _])
+class InvalidPartitionException[P : TypeKey] (val key: PartitionKey[P])
 extends SubdomainException(
   s"partition key $key for ${implicitly[TypeKey[P]].name} has an invalid partition. " +
   s"the properties in the partition must form a prefix of the key value property.")