+++ longevity/src/main/scala/longevity/subdomain/realized/RealizedPartitionKey.scala
@@ -6,7 +6,7 @@ import longevity.subdomain.KeyVal
 import longevity.subdomain.ptype.PartitionKey
 
 private[longevity] case class RealizedPartitionKey[P : TypeKey, V <: KeyVal[P] : TypeKey] private [subdomain](
-  override val key: PartitionKey[P, V],
+  override val key: PartitionKey[P],
   prop: RealizedProp[P, V],
   val partitionProps: Seq[RealizedProp[P, _]],
   val postPartitionProps: Seq[RealizedProp[P, _]],